Cleaned up + fixes

+++ src/routes/post.js
@@ -1,27 +1,17 @@
 const express = require('express');
 const multer = require('multer');
-const admin = require('firebase-admin');
-const { getDownloadURL } = require('firebase-admin/storage');
 const mongoose = require('mongoose');
+
+const firebaseStorage = require('../storages/firebase.storage');
 const requireAuth = require('../middlewares/requireAuth');
-const Post = require('../models/post'); 
-const config = require('../config');
+const Post = require('../models/post');
 const APIError = require('../errors/api.error');
 
-admin.initializeApp({
-    credential: admin.credential.cert(config.fireBase),
-    storageBucket: 'tortuga-backend.appspot.com'
-});
+const fileService = require('../services/file.service');
 
-
-const route = express.Router();
-
-const bucket = admin.storage().bucket();
-
-const basestorage = admin.storage();
 
 const uploadPostPhoto = multer({
-    storage: multer.memoryStorage(),
+    storage: firebaseStorage,
     fileFilter: (req, file, cb) => {
         if (file.mimetype.startsWith('image/')) {
             cb(null, true);
@@ -31,46 +21,32 @@ const uploadPostPhoto = multer({
     }
 });
 
+const route = express.Router();
+
 // New post
-route.post('/', requireAuth(), uploadPostPhoto.fields([{ name: 'photo', maxCount: 1 }]), async (req, res, next) => {
-    if (!req.body.text) {
-        next(new APIError(422, 'Field text is required.'));
+route.post('/', requireAuth(), uploadPostPhoto.single('photo'), async (req, res, next) => {
+    const { title, text } = req.body;
+
+    if (!title || !text) {
+        next(new APIError(422, 'Title/text content is empty!'));
         return;
     }
-    try {
-        const photoFile = req.files['photo'] ? req.files['photo'][0] : null;
-        const photoUrl = photoFile ? await uploadFile(photoFile) : null;
 
-        const { title, text } = req.body;
-        const newPost = new Post({ title, text, photo: photoUrl });
-        await newPost.save();
-        res.send(newPost);
+    try {
+        let file;
+
+        if (req.file) {
+            file = await fileService.saveFile({ fileName: req.file.fileRef.name, firebasePath: req.file.path });
+        }
+
+        const post = await Post.create({ title, text, photo: file ? file.fileName : null });
+
+        res.send(post);
     } catch (error) {
         console.error('Error uploading file or creating post:', error);
         next(new APIError(500, error.message));
     }
 });
-
-// Upload photo
-async function uploadFile(file) {
-    const fileName = Date.now() + '-' + file.originalname;
-    const fileBuffer = file.buffer;
-    const contentType = file.mimetype;
-
-    try {
-        await bucket.file(fileName).save(fileBuffer, {
-            contentType,
-            resumable: false
-        });
-
-        const fileRef = basestorage.bucket().file(fileName);
-        const imageUrl = await getDownloadURL(fileRef);
-        return imageUrl;
-    } catch (error) {
-        console.error('Error uploading file:', error);
-        throw new Error('Error uploading file.');
-    }
-}
 
 // Get all posts
 route.get('/', async (req, res) => {
@@ -91,7 +67,22 @@ route.delete('/:id', requireAuth(), async (req, res) => {
             return res.status(400).json({ message: 'Invalid post ID' });
         }
 
+        const post = await Post.findById(postId);
+        if (!post) {
+            return res.status(404).json({ message: 'Post not found' });
+        }
+
         await Post.findByIdAndDelete(postId);
+
+        // Delete the photo if it exist
+        if (post.photo) {
+            try {
+                await fileService.deleteFile({ fileName: post.photo });
+            } catch (error) {
+                console.error('Error deleting associated photo:', error);
+                return res.status(500).json({ message: 'Failed to delete associated photo' });
+            }
+        }
 
         res.sendStatus(204);
     } catch (error) {
